move lib.rs functions to seperate file
